fibonacci 98

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i;
-	long long int a, b, c;
+	unsigned long int a, b, c;
 
 	a = 1;
 	b = 2;
